refactor: change method parameters to BigDecimal to match entity field type

+++ src/main/java/com/oswaldo/proyecto1/product/repository/ProductRepository.java
@@ -4,11 +4,12 @@ import com.oswaldo.proyecto1.product.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Integer> {
     List<Product> findByNameContainingIgnoreCase(String name);
-    List<Product> findByPriceBetween(double minPrice, double maxPrice);
+    List<Product> findByPriceBetween(BigDecimal minPrice, BigDecimal maxPrice);
     boolean existsByNameAndIdIsNot(String name, Integer productId);
 }
